[Refactor] move classes.py to utils
